usages to unsigned longs

+++ include/process.h
@@ -21,8 +21,8 @@ typedef struct process_info {
     process_state state;
     char name[256];
     char user[64];
-    double cpu_usage;
-    double mem_usage;
+    unsigned long cpu_usage;
+    unsigned long mem_usage;
     long uptime;
 
     unsigned long cpu_utime;
